Popular Android library badge
I'm generating badges to highlight awesome open source Android libraries. They are ranked based on how many top apps the library is used in. If you would like to include it, I've added it to the README. If not, no problem, just close this. Let me know if you have any questions. Cheers!